fix: disable NuGetAudit on template-test
A significant change in NuGet Audit in .NET 9 and VS 17.12 is that the NuGetAuditMode default has changed from “direct” to “all”. Projects using TreatWarningsAsErrors may see new errors (from https://devblogs.microsoft.com/nuget/nugetaudit-2-0-elevating-security-and-trust-in-package-management/).

This was causing the template-test to fail due to a NuGetAudit warning, which was being treated as an error. Since this warning only indicates a dependency package vulnerability -- either direct or indirect -- and does not signify a templating error, NuGetAudit has been disabled.